pyproject.toml
[tool.poetry]
name = "uart"
version = "0.1.0"
description = ""
authors = ["Nikita Saveliev <nikita.saveli3v@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"
pyqt6 = "^6.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
